add manager author

+++ src/module/author/author.service.ts
@@ -1,16 +1,38 @@
 import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { Prisma, Author } from '@prisma/client';
 import { PrismaService } from 'src/Prisma/prisma.service';
-import { createAuthorDto, updateAuthorDto } from './dto/author.dto';
+import { AuthorUpdateDto, createAuthorDto } from './dto/author.dto';
 import { replaceMultipleSpacesAndTrim } from 'src/utils/word';
+import { RoleService } from '../role/role.service';
+import { use } from 'passport';
 
 @Injectable()
 export class AuthorService {
-    constructor(private prisma: PrismaService) { }
+    constructor(private prisma: PrismaService,
+        private readonly roleService: RoleService
+    ) { }
 
-    async findAll(): Promise<Author[]> {
-        return this.prisma.author.findMany();
+    async findAll() {
+        //  tác giả
+        const authors = await this.prisma.author.findMany({
+            include: {
+                NovelAuthor: {
+                    include: {
+                        novel: true,
+                    },
+                },
+            },
+        });
+    
+        // số lượng novel cho từng tác giả
+        const result = authors.map(author => ({
+            ...author,
+            novelCount: author.NovelAuthor.length,
+        }));
+    
+        return result;
     }
+    
 
     async findByName(nickname: string) {
         const author = await this.prisma.author.findMany({
@@ -74,24 +96,21 @@ export class AuthorService {
         return author
     }
 
-    async update(id: number, data: updateAuthorDto): Promise<Author> {
-        const author = await this.findById(id);
-        if (!author) {
-            throw new NotFoundException(`Author with ID ${id} not found`);
-        }
+    async updateAuthor(id: number, data: AuthorUpdateDto) {
         return this.prisma.author.update({
             where: { id },
             data,
         });
     }
 
-    async delete(id: number): Promise<Author> {
-        const author = await this.findById(id);
+    async delete(userId: number, authorId: number) {
+        await this.roleService.checkPermission(userId, 'ManagerAuthor')
+        const author = await this.findById(authorId);
         if (!author) {
-            throw new NotFoundException(`Author with ID ${id} not found`);
+            throw new NotFoundException(`Author with ID ${authorId} not found`);
         }
         return this.prisma.author.delete({
-            where: { id },
+            where: { id: authorId },
         });
     }
